Move project from winforms to Eto
