+++ AM2R ModPacker.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.156
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AM2RModPacker", "AM2RModPacker.csproj", "{7C7450E2-D2AE-4DA3-B8E3-FC953F50A0C7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AM2RModPacker", "AM2RModPacker\AM2RModPacker.csproj", "{13E58D3E-D34E-4354-B5D3-AC43B05E7212}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7C7450E2-D2AE-4DA3-B8E3-FC953F50A0C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C7450E2-D2AE-4DA3-B8E3-FC953F50A0C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C7450E2-D2AE-4DA3-B8E3-FC953F50A0C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C7450E2-D2AE-4DA3-B8E3-FC953F50A0C7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{13E58D3E-D34E-4354-B5D3-AC43B05E7212}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{13E58D3E-D34E-4354-B5D3-AC43B05E7212}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{13E58D3E-D34E-4354-B5D3-AC43B05E7212}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{13E58D3E-D34E-4354-B5D3-AC43B05E7212}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
