+++ AM2RModPacker/Info.plist
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+  <dict>
+    <key>CFBundleName</key>
+    <string>AM2RModPacker</string>
+    <key>CFBundleIdentifier</key>
+    <string>com.example.AM2RModPacker</string>
+    <key>CFBundleShortVersionString</key>
+    <string>1.0</string>
+    <key>LSMinimumSystemVersion</key>
+    <string>10.14</string>
+    <key>CFBundleDevelopmentRegion</key>
+    <string>en</string>
+    <key>NSHumanReadableCopyright</key>
+    <string></string>
+    <key>CFBundleIconFile</key>
+    <string>MacIcon.icns</string>
+  </dict>
+</plist>
