+++ AM2RModPacker/utilities/android/LICENSE
