--- packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.WindowsAPICodePack-Core" version="1.1.0.0" targetFramework="net472" />
-  <package id="Microsoft.WindowsAPICodePack-Shell" version="1.1.0.0" targetFramework="net472" />
-  <package id="WindowsAPICodePack-Core" version="1.1.1" targetFramework="net472" />
-  <package id="WindowsAPICodePack-Shell" version="1.1.1" targetFramework="net472" />
-</packages>
